+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Refactor user command and database query formatting. (`cb17155`)
 - Add user onboarding invitation management functionality. (`801d6f5`)
 - Update Dockerfile to install ca-certificates along with curl. (`f47adf6`)